+++ R/RECOM_RANDOM.R
@@ -3,7 +3,7 @@
 RANDOM <- function(data = NULL, parameter = NULL) {
 
   model <- list(
-    range = range(getRatings(data), na.rm = TRUE),
+    range = NA,
     labels = colnames(data)
   )
 
@@ -26,6 +26,8 @@ RANDOM <- function(data = NULL, parameter = NULL) {
     }
 
     if(ncol(newdata) != length(model$labels)) stop("number of items in newdata does not match model.")
+
+    ### FIXME: topN lists could be created directly more memory efficient.
 
     ratings <- new("realRatingMatrix",
       data = dropNA(matrix(runif(n = nrow(newdata)*ncol(newdata),