+++ R/topNList.R
@@ -62,35 +62,36 @@ setMethod("getTopNLists", signature(x = "realRatingMatrix"),
     # just in case
     x <- denormalize(x)
 
-    x.l <- getList(x, decode = FALSE)
+    ### FIXME: this takes a lot of memory for a dense rating matrix
+    reclist <- getList(x, decode = FALSE)
 
     if(is.null(randomize) || is.na(randomize)) {
-      reclist <- lapply(x.l, FUN = function(l)
+      reclist <- lapply(reclist, FUN = function(l)
         head(sort(l, decreasing = TRUE), n=n))
 
-      ret <- new("topNList",
+      reclist <- new("topNList",
         items = lapply(reclist, FUN = function(l) as.integer(names(l))),
         ratings = lapply(reclist, as.vector),
         itemLabels = colnames(x), n = n)
 
       if(!is.null(minRating) && !is.na(minRating))
-        ret <- bestN(ret, n = n, minRating = minRating)
+        reclist <- bestN(reclist, n = n, minRating = minRating)
 
     }else{
     ## randomize recommendations
-      reclist <- lapply(x.l, FUN = function(l) {
+      reclist <- lapply(reclist, FUN = function(l) {
         if(!is.null(minRating) && !is.na(minRating)) l <- l[l>=minRating]
         if(length(l)>0) sample(l, size = min(n, length(l)),
           prob = (l-min(l)+1)^randomize)
         else integer(0)
       })
 
-      ret <- new("topNList",
+      reclist <- new("topNList",
         items = lapply(reclist, FUN = function(l) as.integer(names(l))),
         ratings = lapply(reclist, as.vector),
         itemLabels = colnames(x), n = n)
     }
-  ret
+  reclist
   })
 
 ## only keep best n items.